make  jest configuration files for different environments
