+++ jest.config.js
@@ -1,43 +1,6 @@
-// // this is for Prisma ORM unit testing:
-// /** @type {import('ts-jest').JestConfigWithTsJest} */
-// module.exports = {
-//   preset: "ts-jest",
-//   testEnvironment: "node",
-//   setupFilesAfterEnv: ["<rootDir>/singleton.ts"],
-// };
-
-//// ######
-
-// //import type { Config } from "jest";
-const nextJest = require("next/jest");
-
-const createJestConfig = nextJest({
-  // Provide the path to your Next.js app to load next.config.js and .env files in your test environment
-  dir: "./",
-});
-
-// Add any custom config to be passed to Jest
-/** @type {import('ts-jest').JestConfigWithTsJest} */
+/** @type {import('jest').Config} */
 const config = {
-  // Add more setup options before each test is run
-  setupFilesAfterEnv: ["<rootDir>/jest.setup.ts"],
-  testEnvironment: "jest-environment-jsdom",
-  preset: "ts-jest",
+  projects: ["<rootDir>/tests/node", "<rootDir>/tests/jsdom"],
 };
 
-// createJestConfig is exported this way to ensure that next/jest can load the Next.js config which is async
-module.exports = createJestConfig(config);
-
-// // ######
-
-// /** @type {import('ts-jest').JestConfigWithTsJest} */
-// const nextJest = require("next/jest");
-// const createJestConfig = nextJest({
-//   // Provide the path to your Next.js app to load next.config.js and .env files in your test environment
-//   dir: "./",
-// });
-// const config = {
-//   verbose: true,
-// };
-
-// module.exports = createJestConfig(config);
+module.exports = config;
